requirements.txt
numpy~=1.21.2
pygame~=2.0.1
tensorflow-gpu==2.7.0